Re-work Song object properties
For ease of implementation the CamelCase properties were used
since they mapped directly to the JSON keys exported by TinySong.
As of now the CamelCase properties and the new properties will live
alongside one another with the new-style properties being preferred
in tests and docs.

This commit adds a map that translates the JSON data to the new-style
properties and an additional test to ensure that both the old-style
properties and new-style properties stay happy.